+++ NEWS.md
@@ -1,6 +1,7 @@
 # tna 0.4.4
 
-* The `plot.tna_centralities` and `plot.group_tna_centralities` functions now plot the centralities in the same order as provided in the `measures` argument.
+* The `plot.tna_centralities()` and `plot.group_tna_centralities()` functions now plot the centralities in the same order as provided in the `measures` argument.
+* The `plot.tna()` and `plot_model()` functions now use the median edge weight as the default value for the `cut` argument.
 
 # tna 0.4.3